+++ main/output.c
@@ -18,7 +18,7 @@
    +----------------------------------------------------------------------+
 */
 
-/* $Id: output.c,v 1.148 2003/02/09 19:09:52 sniper Exp $ */
+/* $Id: output.c,v 1.149 2003/03/06 23:07:28 ddhill Exp $ */
 
 #include "php.h"
 #include "ext/standard/head.h"
@@ -712,7 +712,7 @@ static int php_ub_body_write(const char *str, uint str_length TSRMLS_DC)
 PHP_FUNCTION(ob_start)
 {
 	zval *output_handler=NULL;
-	uint chunk_size=0;
+	long chunk_size=0;
 	zend_bool erase=1;
 	int argc = ZEND_NUM_ARGS();
